third_party/blink/web_tests/fast/xsl/dtd-in-source-document-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
This tests that entities defined in an external DTD referenced to from the source document are replaced correctly. If this test is successful, the text SUCCESS should be shown below.
SUCCESS
